requirements.txt
fastapi
uvicorn
firebase-admin
llama-index-core
google-genai
sentence-transformers
rank_bm25
python-dotenv
numpy
pandas
selenium
